refactor(frontend): extract admin nav item to a shared module
Move NavExtensionItem, AdminIcon, and the {to, label, icon} payload into
a new frontend/src/extensions/admin-nav-item.ts that exports an
ADMIN_NAV_ITEM constant. nav.ts now re-exports the type and returns
[ADMIN_NAV_ITEM] from getExtraNavItems.

This lets the premium overlay's own nav.ts import ADMIN_NAV_ITEM
instead of carrying its own copy of the icon and the interface (a
duplication that CLAUDE.md's "no code duplication between OSS and
premium" rule calls out, and the one that caused issue #1357 to also
slip through unfixed on the premium side after the OSS-only fix in #1360).

Existing nav test still exercises getExtraNavItems; an additional pair
of tests pins the shape and rendered icon of ADMIN_NAV_ITEM directly.
