+++ frontend/src/extensions/admin-nav-item.ts
@@ -1,0 +1,34 @@
+import { createElement, type ComponentType } from 'react';
+
+export interface NavExtensionItem {
+  to: string;
+  label: string;
+  icon: ComponentType;
+  /** Fired in addition to navigation, so extensions can reset internal state (e.g. hash-routed tabs). */
+  onClick?: () => void;
+}
+
+function AdminIcon() {
+  return createElement(
+    'svg',
+    {
+      className: 'w-5 h-5 shrink-0',
+      fill: 'none',
+      stroke: 'currentColor',
+      viewBox: '0 0 24 24',
+    },
+    createElement('path', {
+      strokeLinecap: 'round',
+      strokeLinejoin: 'round',
+      strokeWidth: 1.5,
+      d: 'M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z',
+    }),
+  );
+}
+
+/** Shared admin nav item. Premium imports this and adds an isAdmin gate. */
+export const ADMIN_NAV_ITEM: NavExtensionItem = {
+  to: '/app/admin',
+  label: 'Admin',
+  icon: AdminIcon,
+};
